sim/knxsim/requirements.txt
# KNX Simulator + Management API
pyyaml>=6.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
